+++ changelog.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Table Iteration ends after the second iteration ([Issue #70](https://github.com/michaeldcanady/servicenow-sdk-go/issues/70))
+- `TableRequestBuilder.Post`/`TableRequestBuilder.Post2` sending `PUT` request ([Issue #69](https://github.com/michaeldcanady/servicenow-sdk-go/issues/69))
 
 ## [1.2.1] - 10/12/2023